ADD test for adoption

+++ src/controllers/pets.controller.js
@@ -11,20 +11,25 @@ const getAllPets = async (req, res) => {
 const createPet = async (req, res, next) => {
   const { name, specie, birthDate } = req.body;
   if (!name || !specie || !birthDate) return next(BadRequest);
+
   const pet = PetDTO.getPetInputFrom({ name, specie, birthDate });
   const result = await petsService.create(pet);
   res.send({ status: "success", payload: result });
 };
 
-const updatePet = async (req, res) => {
+const updatePet = async (req, res, next) => {
   const petUpdateBody = req.body;
   const petId = req.params.pid;
+  if (!petId || !petUpdateBody) return next(BadRequest);
+
   await petsService.update(petId, petUpdateBody);
   res.send({ status: "success", message: "pet updated" });
 };
 
-const deletePet = async (req, res) => {
+const deletePet = async (req, res, next) => {
   const petId = req.params.pid;
+  if (!petId) return next(BadRequest);
+
   await petsService.delete(petId);
   res.send({ status: "success", message: "pet deleted" });
 };
@@ -33,14 +38,14 @@ const createPetWithImage = async (req, res, next) => {
   const file = req.file;
   const { name, specie, birthDate } = req.body;
   if (!name || !specie || !birthDate) return next(BadRequest);
-  console.log(file);
+
   const pet = PetDTO.getPetInputFrom({
     name,
     specie,
     birthDate,
     image: `${__dirname}/../public/img/${file.filename}`,
   });
-  console.log(pet);
+
   const result = await petsService.create(pet);
   res.send({ status: "success", payload: result });
 };
